backend/user/auth/requirements.txt
blinker==1.9.0
boto3==1.35.69
botocore==1.35.69
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
PyJWT==2.10.0
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.3
